+++ 3_poe/poe_node/runtime/src/lib.rs
@@ -138,7 +138,7 @@ parameter_types! {
 	pub BlockLength: frame_system::limits::BlockLength = frame_system::limits::BlockLength
 		::max_with_normal_ratio(5 * 1024 * 1024, NORMAL_DISPATCH_RATIO);
 	pub const SS58Prefix: u8 = 42;
-	pub const ClaimLengthLimit: u32 = 100;
+	pub const ClaimLengthLimit: usize = 100;
 }
 
 // Configure FRAME pallets to include in runtime.
